int replace by Long

+++ bookStore/src/main/java/com/bookStore/entity/Books.java
@@ -13,8 +13,8 @@ public class Books {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
+	private Long id;
+	 
 	private String name;
 	
 	private String author;
